study2/requirements.txt
anthropic
openai
google-genai
yfinance
pandas
numpy
matplotlib
seaborn
scipy
pyyaml
python-dotenv
rich
